Update Code.v
Introduction: Welcome to the first day of our 60-day Verilog challenge! Today, we’ll start with a simple but fundamental task to get you familiar with Verilog syntax and different coding styles.

Task:  In this challenge, we explored different ways to create a Verilog module that outputs a constant high signal.

Why This Matters: This exercise helps you understand the flexibility and various approaches you can take when writing Verilog code. It also sets the foundation for more complex designs in future challenges.